refactor: create utils module and move VectorEngine and helper functions into utils
