+++ test/test_ModelPack.py
@@ -5,7 +5,7 @@ class TestModelPack(unittest.TestCase):
     '''
     Run tests on ModelPack class in models.py
     '''
-    # Test case 1: handle invalid source input
+    # Test case: handle invalid source input
     def test_invalid_source_input(self):
         # test invalid source as int
         with self.assertRaises(ValueError):
@@ -17,7 +17,7 @@ class TestModelPack(unittest.TestCase):
         with self.assertRaises(ValueError):
             m = ModelPack(model='gpt2', source='huggingface1')
 
-    # Test case 2: handle invalid model input
+    # Test case: handle invalid model input
     def test_invalid_model_input(self):
         # test invalid model as int
         with self.assertRaises(ValueError):
@@ -29,8 +29,8 @@ class TestModelPack(unittest.TestCase):
         with self.assertRaises(ValueError):
             m = ModelPack(model='gpt99', source='huggingface')
 
-    # Test case 3: handle model and source mismatch input
-    def test_invalid_model_source_input(self):
+    # Test case: handle model and source mismatch input
+    def test_incorrect_model_source_input(self):
         # test invalid model and source match combo 1
         with self.assertRaises(ValueError):
             m = ModelPack(model='text-davinci-002', source='huggingface')
@@ -38,19 +38,19 @@ class TestModelPack(unittest.TestCase):
         with self.assertRaises(ValueError):
             m = ModelPack(model='gpt2', source='openai')
 
-    # Test case 4: handle model and source correct match input
-    def test_correct_model_source_input(self):
+    # Test case: handle model and source correct match input
+    def test_correct_model_source_gpt2_input(self):
         # test valid model and source match combo 1
         m = ModelPack(model='gpt2', source='huggingface')
 
-    # Test case 5: handle no OpenAI API key input
-    def test_no_openai_api_key_input(self):
+    # Test case: handle no OpenAI API key input
+    def test_missing_openai_api_key_input(self):
         # test no openai api key provided
         with self.assertRaises(ValueError):
             m = ModelPack(model='text-davinci-002', source='openai')
     
-    # Test case 6: handle model and source correct match input
-    def test_correct_model_source_input(self):
+    # Test case: handle model and source correct match input
+    def test_correct_model_source_flan_input(self):
         # test valid model and source match combo 2
         m = ModelPack(model='google/flan-t5-small', source='huggingface')
